requirements/test.txt
# Test dependencies go here.
-r base.txt

# coverage==4.0.2
# django_coverage_plugin==1.0
flake8==2.5.0
django-test-plus==1.0.11
fake-factory==0.5.3
factory_boy==2.6.0
